rename test module; tidy up w/ fixutres and parameterize
